+++ fiddles/bash/fiddle-0100-ReactCreateApp/scripts/bin/react/_update.sh
@@ -102,8 +102,6 @@ function update() {
     rmNodeModules || exit $?;
     rmPackageLock || exit $?;
     npmCheckUpdates || exit $?;
-    revertTypeScript || exit $?;
-    updateAngularCli || exit $?;
     npmShrinkWrap || exit $?;
   )