+++ fiddles/bash/fiddle-0100-ReactCreateApp/CHANGELOG.md
@@ -5,4 +5,6 @@
 * Added [fiddles/react/fiddle-0100-Fork](fiddles/react/fiddle-0100-Fork)
 * Added [fiddles/react/fiddle-0100-Fork](fiddles/react/fiddle-0100-Fork)
 * Added [fiddles/react/fiddle-0100-Fork](fiddles/react/fiddle-0100-Fork)
-* Added [fiddles/react/fiddle-0001-Fork](fiddles/react/fiddle-0001-Fork)
+* Added [fiddles/react/fiddle-0100-Fork](fiddles/react/fiddle-0100-Fork)
+* Added [fiddles/react/fiddle-0100-Fork](fiddles/react/fiddle-0100-Fork)
+* Added [fiddles/react/fiddle-0100-Fork](fiddles/react/fiddle-0100-Fork)